refactor: modularize analytics and add build tooling
