+++ new_site (1)/analytics.html
@@ -94,24 +94,6 @@
   </footer>
 
   <!-- Data and analytics scripts -->
-  <!-- Leaflet library for map rendering -->
-  <script
-    src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js"
-    integrity="sha256-20nQCchB9co0qIjJZRGuk2/Z9VM+kNiyxNV1lvTlZBo="
-    crossorigin="anonymous"
-  ></script>
-  <!-- Leaflet MarkerCluster for clustering functionality -->
-  <script
-    src="https://unpkg.com/leaflet.markercluster@1.4.1/dist/leaflet.markercluster.js"
-    integrity="sha256-WL6HHfYfbFEkZOFdsJQeY7lJG/E5airjvqbznghUzRw="
-    crossorigin="anonymous"
-  ></script>
-  <!-- Leaflet.heat for heatmap functionality -->
-  <script
-    src="https://unpkg.com/leaflet.heat@0.2.0/dist/leaflet-heat.js"
-    integrity="sha256-65UqrlgGoRAnKfKRuriH3eeDrOhZgZo1SCenduc+SGo="
-    crossorigin="anonymous"
-  ></script>
-  <script src="analytics.js"></script>
+  <script type="module" src="analytics.js"></script>
 </body>
 </html>
