+++ new_site (1)/vite.config.js
@@ -1,0 +1,15 @@
+import { defineConfig } from 'vite';
+
+export default defineConfig({
+  root: '.',
+  build: {
+    rollupOptions: {
+      input: {
+        main: 'index.html',
+        analytics: 'analytics.html',
+        data: 'data.html',
+        blog: 'blog.html'
+      }
+    }
+  }
+});
